Refactor kubo tests.
A new function check_integrate_dynamical, which is similar to check_integrate, is defined.
The differences between them are 1) the shape of the data are different for dynamical quantities,
and 2) opt_conductivity requires a special treatment because of sym and asym data.
I defined a new function to not make the functions too complicated.