Modification of F matrix defined in the ab initio mesh
get_FF_qb() was used to compute the real-space matrix 'OO'
involved in the Berry curvature from uIu, to test the
implementation of uIu. Such real-space matrix can be directly
obtained from mmn, as it is already implemented. For this
reason get_FF_qb() has been eliminated, but it can always be
recovered for testing.

get_FF_qb_tot() was used to get the full F matrix, with its
symmetric and antisymmetric parts. It has been replaced by
get_GG_qb(), which focuses on the symmetric part only. This is
the only piece of the F matrix that unavoidably requires the
uIu matrix elements.

Similar changes in system_w90.py

Also minor changes for Lint / flake test