Subsequent modifications in system.py, and uncommented set_wannier_centers()
The version of Jae-Mo's approach we have implemented follows
phase convention I (it requires use_wcc_phase=True). It does not
require adding or subtracting quantities to the real-space
matrices 'AA', 'BB', 'CC' in this phase convention, so this
procedure in set_wannier_centers() is skipped if
transl_inv_JM=True.

Also minor modifications associated to 'FF' and 'GG'.