Implementation of the intrinsic multipole moments as lazy properties
The intrinsic multipole moments are defined as E1, M1, and E2. In
addition, Vn is defined for the band velocity, Bln for the
combination of M1 and E2 that appears in the spatially-dispersive
conductivity tensor, and kron to identify for a given k point
which bands are degenerate for a given energy threshold.

Levi-Civita symbol added to __utility.py for the computation of
Bln.