Addition of matrix OO to system for testing
The OO matrix (Berry curvature) is currently computed through
the matrix elements of the position operator. It is the way it is
computed when 'Berry = True'. It can be alternatively computed through
the anti-symmetric piece of the real-space matrix obtained from uIu. I
have added OO to system, which is activated through 'OSD = True', to
simplify the comparison between these two approaches.