Spatially-dispersive conductivity tensor (#308)
*translationally invariant finite-difference schemes including Jae-Mo's method extended to other real-space matrix elements.


* Modification of F matrix defined in the ab initio mesh

get_FF_qb() was used to compute the real-space matrix 'OO'
involved in the Berry curvature from uIu, to test the
implementation of uIu. Such real-space matrix can be directly
obtained from mmn, as it is already implemented. For this
reason get_FF_qb() has been eliminated, but it can always be
recovered for testing.

get_FF_qb_tot() was used to get the full F matrix, with its
symmetric and antisymmetric parts. It has been replaced by
get_GG_qb(), which focuses on the symmetric part only. This is
the only piece of the F matrix that unavoidably requires the
uIu matrix elements.


* Subsequent modifications in system.py, and uncommented set_wannier_centers()

The version of Jae-Mo's approach we have implemented follows
phase convention I (it requires use_wcc_phase=True). It does not
require adding or subtracting quantities to the real-space
matrices 'AA', 'BB', 'CC' in this phase convention, so this
procedure in set_wannier_centers() is skipped if
transl_inv_JM=True.

Also minor modifications associated to 'FF' and 'GG'.

* Implementation of the intrinsic multipole moments as lazy properties

The intrinsic multipole moments are defined as E1, M1, and E2. In
addition, Vn is defined for the band velocity, Bln for the
combination of M1 and E2 that appears in the spatially-dispersive
conductivity tensor, and kron to identify for a given k point
which bands are degenerate for a given energy threshold.

Levi-Civita symbol added to __utility.py for the computation of
Bln.

* Implementation of the spatially-dispersive conductivity tensor

The spatially-dispersive conductivity tensor has been implemented
as a DynamicCalculator. Only Fermi sea contributions are complete.

Changes in formula/covariant.py undone, it is not necessary to
implement these formulas there right now.

* Minor changes undone for better comparison with master branch

* Solving comments on pull request #1

* LeviCivita array eliminated from __utility.py, the
antisymmetric B matrix in data_K.py can be obtained from
alpha_A and beta_A.

* kron() function in data_K.py simplified.

* Following Jae-Mo's approach, the phase factors are computed
only once. Also many for loops were eliminated for better
readability.

* Minor edit on covariant.py

Irrelevant blank line was added by accident in a previous commit.
Now it is eliminated.

* Implementation of SDCT

* Fermi surface contribution included in dynamic calculators.
The derivative of the Fermi-Dirac distribution function is
introduced analytically for a finite temperature kBT.

* Main calculators SDCT_asym() and SDCT_sym(), which sum the
corresponding two Fermi sea and two Fermi surface sub-calculators
for the spatially-dispersive conductivity tensor. It was
necessary to modify __add function in _energyresult.py.

* Addition of matrix OO to system for testing

The OO matrix (Berry curvature) is currently computed through
the matrix elements of the position operator. It is the way it is
computed when 'Berry = True'. It can be alternatively computed through
the anti-symmetric piece of the real-space matrix obtained from uIu. I
have added OO to system, which is activated through 'OSD = True', to
simplify the comparison between these two approaches.

* Minor change of Fermi-surface contributions in dynamic.py

This is a more convenient and proper way to specify the Fermi factor.

* Test on real-space matrix elements using Jae-Mo's approach

The test is performed for Silicon, with four valence bands
wannierized. Phase convention I is assumed.

* Minor edits on test of Jae-Mo's approach

Eliminate unnecessary files, and corrected lines for flake8.

* Minor edits on test of Jae-Mo's approach

Eliminate unnecessary files, and corrected lines for flake8.

* Multipole decomposition of SDCT (antisymmetric)

There are now three flags for the antisymmetric piece of the
spatially-dispersive conductivity tensor: M1 (magnetic dipole),
E2 (electric quadrupole), and V (band dispersive). Their
default value is True. To select only one of them, include the
rest with value False in the arguments of the calculator.

* Several updates and bug corrected in SDCT

** Flag to distinguish 'internal' and 'external' contributions.

** Spin contribution to the intrinsic magnetic dipole moment.

** Imaginary term of the frequency (eta) transformed to
'smr_fixed_width'.

** Bug corrected regarding 'fermi_sea_I' calculators of the
spatially-dispersive conductivity tensor: swapaxes() was
missing for the B matrix.

* Updates on the symmetric piece of SDCT

** Bug corrected: A factor 2 was missing in Formula_SDCT_sym_sea_II().

** Distinction between internal and external terms for the electric dipole
moment.

* nspin=2 case in interface PythTB -> WB

Incorporate the interface with nspin=2 according to the following PR:

https://github.com/wannier-berri/wannier-berri/pull/299

* fixed some errors, temporarily weakened flake

* fixed bug in tb_py

* fixed old tests, the new Si test is failing

* update data for CC in Si_JM, softer accuracy for disentangle

* KaneMele model and test, spin=True for PythTB

* lint

* test for SDCT

* updated OO.npz for Si

* lowered precision for some terms

* lowered precision for some terms

* added Chiral_OSD and tests

* Revised recentering term in CC

* update  reference data

* test SDCT on random system

* fixed pep issues

* test for wcc_phase_fin_diff

---------
Co-authored-by: Seung-Ju Hong <sjhong6230@snu.ac.kr>
Co-authored-by: Stepan Tsirkin <stepan.tsirkin@ehu.eus>